current code
has all the games in it. Both minesweeper and guess the number need some type of back number implemented to get back to the main screen. The main screen has been changed so that when a game is opened it is closed. tic tac toe and rps were updated to go back to the main screen after the quit button is clicked.